标签页面，标签、目录添加、删除、搜索接口

+++ blog-api/src/main/java/com/erotsx/blog/dao/ArticleMapper.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface ArticleMapper extends BaseMapper<Article> {
 
-    List<Integer> findArticleIdsByTagId(Integer tagId);
+    List<Long> findArticleIdsByTagId(Long tagId);
 }
