+++ blog-api/src/main/java/com/erotsx/blog/utils/JWTUtils.java
@@ -13,7 +13,7 @@ public class JWTUtils {
 
     private static final String jwtToken = "123456Mszlu!@###$$";
 
-    public static String createToken(int userId) {
+    public static String createToken(Long userId) {
         Map<String, Object> claims = new HashMap<>();
         claims.put("userId", userId);
         JwtBuilder jwtBuilder = Jwts.builder()
